% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EllDistrEstDeriv.R
\name{compute_etahat}
\alias{compute_etahat}
\title{Compute \eqn{\hat{\eta}_k}}
\usage{
compute_etahat(
  X,
  mu = 0,
  Sigma_m1 = diag(d),
  grid,
  h,
  Kernel = "gaussian",
  a = 1,
  k,
  mpfr = FALSE,
  precBits = 100,
  dopb = TRUE
)
}
\arguments{
\item{k}{order of the derivative}
}
\value{
a vector of size \code{n1 = length(grid)}.
Each component of this vector is \eqn{\hat{\eta}_k(x[i])}
where \code{x[i]} is the \eqn{i}-th element of the grid.
}
\description{
\eqn{\hat{\eta}_k} is a quantity that is useful
for estimating the \eqn{k}-th derivative of the generator
of an elliptical distribution.
}
\examples{

# Comparison between the estimated and true generator of the Gaussian distribution
n = 500000
d = 3
X = matrix(rnorm(n * d), ncol = d)
grid = seq(0, 5, by = 0.1)
a = 0.7

etahat = compute_etahat(X = X, grid = grid, a = a, h = 0.04, k = 1)
plot(grid, etahat, type = "l", ylim = c(-0.02, 0.02))

# Computation of true values
g = exp(-grid/2)/(2*pi)^{3/2}
gprime = (-1/2) *exp(-grid/2)/(2*pi)^{3/2}
A = a^(d/2)
psia = -a + (A + grid^(d/2))^(2/d)
psiaprime = grid^(d/2 - 1) * (A + grid^(d/2))^(2/d - 1)
psiasecond = psiaprime * ( (d-2)/2 ) * grid^{-1} * A *
  ( grid^(d/2) + A )^(-1)

rhoprimexi = ((d-2) * grid^((d-4)/2) * psiaprime
- 2 * grid^((d-2)/2) * psiasecond) / (2 * psiaprime^3) * g +
grid^((d-2)/2) / (psiaprime^2) * gprime

lines(grid, rhoprimexi, col = "red")

}
\author{
Victor Ryan, Alexis Derumigny
}
