% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_auxiliaryFunctions.R
\name{compute_matrix_alpha}
\alias{compute_matrix_alpha}
\title{Compute the matrix of coefficients alpha}
\usage{
compute_matrix_alpha(kmax, grid, a, d)
}
\arguments{
\item{kmax}{order the derivative that we want to compute}

\item{grid}{the grid of the values at which we want to compute the derivative}

\item{a}{the tuning parameter controlling the bias of the estimator
at zero.}

\item{d}{the dimension of the problem}
}
\value{
a \code{(kmax+1) * (kmax+1) * length(grid)} array
}
\description{
Compute the matrix of coefficients alpha
}
\examples{
kmax = 1
d = 3
grid = 0.2
a = 0.8
compute_matrix_alpha(kmax = kmax, grid = grid, a = a, d = d)

}
\references{
Ryan, V., & Derumigny, A. (2024).
On the choice of the two tuning parameters for nonparametric estimation of an
elliptical distribution generator
\href{https://arxiv.org/abs/2408.17087}{arxiv:2408.17087}.
}
\author{
Victor Ryan, Alexis Derumigny
}
\keyword{internal}
