% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondEllGenEst.R
\name{CondEllGenEst}
\alias{CondEllGenEst}
\title{Conditional Density Generator Estimator for Elliptical Distributions}
\usage{
CondEllGenEst(
  dataMatrix,
  observedZ,
  mu,
  sigma,
  gridZ,
  grid,
  h,
  Kernel = "epanechnikov",
  a = 1,
  h_psi = NULL,
  mpfr = FALSE,
  precBits = 100,
  dopb = TRUE
)
}
\arguments{
\item{dataMatrix}{a matrix of size \eqn{n \times d} containing the \eqn{n}
observations of the \eqn{d}-dimensional response variable \eqn{X}. The pairs
\eqn{(X_i, Z_i)} are assumed to be i.i.d. realizations of a joint random vector.}

\item{observedZ}{vector with \eqn{n} observations of the conditioning
variable \eqn{Z}.}

\item{mu}{matrix of size \eqn{d \times \code{length(gridZ)}} with
conditional means \eqn{\mu(z)}.}

\item{sigma}{array of size \eqn{d \times d \times
\code{length(gridZ)}} with conditional covariances \eqn{\Sigma(z)}.}

\item{gridZ}{vector of points \eqn{z} at which the conditional covariance matrix
\eqn{\Sigma(z) = \mathrm{Cov}(X \mid Z = z)} is estimated.}

\item{grid}{vector of \eqn{\xi} values where the generator is
evaluated.}

\item{h}{bandwidth of the kernel for kernel smoothing over \eqn{Z}.}

\item{Kernel}{name of the kernel. Possible choices are
\code{"gaussian"}, \code{"epanechnikov"}, \code{"triangular"}.}

\item{a}{tuning parameter controlling bias at \eqn{\xi = 0}.}

\item{h_psi}{optional bandwidth for kernel smoothing over
\eqn{\psi(\xi)}. If not specified, the Silverman's rule of thumb is used.}

\item{mpfr}{if \code{mpfr = TRUE}, multiple precision floating point is used
via the package \link[Rmpfr:Rmpfr]{Rmpfr}.
This allows for a higher (numerical) accuracy, at the expense of computing time.
It is recommended to use this option for higher dimensions.}

\item{precBits}{number of precBits used for floating point precision
(only used if \code{mpfr = TRUE}).}

\item{dopb}{a Boolean value.
If \code{dopb = TRUE}, a progress bar is displayed.}
}
\value{
A matrix of size \eqn{\code{length(grid)} \times
  \code{length(gridZ)}}, \eqn{\widehat{g}_{n}(\xi \mid z)} containing the
estimated conditional density generator at each \eqn{\xi} and \eqn{z}.
}
\description{
This function estimates the conditional density generator \eqn{g_z} of a
continuous elliptical distribution. For a \eqn{d}-dimensional response
\eqn{X} conditional on \eqn{Z = z}, the conditional density is of the form
\deqn{
f_{X|Z}(x \mid z) = |\Sigma(z)|^{-1/2} \,
g_z\left( (x - \mu(z))^\top \, \Sigma(z)^{-1} \, (x - \mu(z)) \right),
}
where \eqn{x \in \mathbb{R}^d}, \eqn{z \in \mathbb{R}}, \eqn{\mu(z) \in
\mathbb{R}^d} is the conditional mean, \eqn{\Sigma(z)} is the \eqn{d \times d}
conditional covariance matrix, and \eqn{g_z: \mathbb{R}_+ \rightarrow
\mathbb{R}_+} is the \strong{conditional density generator}.
}
\details{
Given a sample \eqn{(X_1, Z_1), \dots, (X_n, Z_n)}, the conditional density
generator at a point \eqn{\xi} and covariate value \eqn{z} is estimated by
\deqn{
\widehat{g}_{n}(\xi \mid z)
:= \frac{\xi^{1-d/2} \, \psi'(\xi)}{s_d \, n h_{\psi}}
\sum_{i=1}^n w_{n,i}(z)
\left[
  K\left( \frac{\psi_a(\xi) - \psi_a(\xi_i(z))}{h_{\psi}} \right)
+ K\left( \frac{\psi_a(\xi) + \psi_a(\xi_i(z))}{h_{\psi}} \right)
\right],
}
where
\deqn{s_d := \pi^{d/2} / \Gamma(d/2), \quad
\psi(\xi) := -a + (a^{d/2} + \xi^{d/2})^{2/d}, \quad
\xi_i(z) := (X_i - \mu(z))^\top \Sigma(z)^{-1} (X_i - \mu(z)),
}
\eqn{h > 0} and \eqn{a > 0} are tuning parameters, \eqn{K} is a kernel
function, and \eqn{w_{n,i}(z)} are Nadaraya-Watson weights:
\deqn{
w_{n,i}(z) = \frac{K\left( (z - Z_i)/h \right)}{
\sum_{j=1}^n K\left( (z - Z_j)/h \right)}.
}
}
\examples{
# Conditional generator with Z-dependent mean/covariance

n  = 5000
d  = 3
Z  = runif(n)

data_X = matrix(NA, n, d)
for (i in 1:n) {
  mean_i  = rep(Z[i], d)
  sigma_i = matrix(c(Z[i],   Z[i]/2, Z[i]/3,
                     Z[i]/2, Z[i],   Z[i]/2,
                     Z[i]/3, Z[i]/2, Z[i]), nrow = d)
  data_X[i,] = EllDistrSim(
    n = 1, d = d,
    A = chol(sigma_i),
    mu = mean_i,
    density_R2 = function(x) { (2*pi)^(-d/2) * exp(-x/2) * (x > 0) }
  )
}

h     <- 1.06 * sd(Z) * n^(-1/5)
gridZ <- c(0.2, 0.8)

mu_est    = CondMeanEst(data_X, Z, gridZ, h)
Sigma_est = CondCovEst(data_X, Z, gridZ, h, type = "grid_mean")

grid = seq(0, 8, length.out = 80)
g_est  = CondEllGenEst(
  dataMatrix = data_X,
  observedZ  = Z,
  mu         = mu_est,
  sigma      = Sigma_est,
  gridZ      = gridZ,
  grid       = grid,
  h          = h,
  Kernel     = "epanechnikov",
  a          = 1
)

g_true <- function(r){ (2*pi)^(-d/2) * exp(-r/2) }
g_grid_true = g_true(grid)

plot(grid, g_est[,1], type="l", col="blue", lwd=2,
  main="Conditional Generator: Estimated vs True",
  xlab="u", ylab="g(u | Z)")
  lines(grid, g_est[,2], col="red", lwd=2)
  lines(grid, g_grid_true, col="black", lwd=2, lty=2)
  legend("topright",
  legend=c("Estimated Z = 0.2", "Estimated Z = 0.8", "True Gaussian generator"),
  col=c("blue", "red", "black"), lwd=2, lty=c(1,1,2))

}
\references{
Liebscher, E. (2005). A semiparametric density estimator based on
elliptical distributions. Journal of Multivariate Analysis, 92(1), 205-222.
}
