% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondMeanEst.R
\name{CondMeanEst}
\alias{CondMeanEst}
\title{Conditional Mean Estimator}
\usage{
CondMeanEst(dataMatrix, observedZ, gridZ, h, Kernel = "epanechnikov")
}
\arguments{
\item{dataMatrix}{a matrix of size \eqn{n \times d} containing the \eqn{n}
observations of the \eqn{d}-dimensional response variable \eqn{X}. The pairs
\eqn{(X_i, Z_i)} are assumed to be i.i.d. realizations of a joint random vector.}

\item{observedZ}{vector with \eqn{n} observations of the conditioning
variable \eqn{Z}.}

\item{gridZ}{vector of points \eqn{z} at which the conditional mean
\eqn{\mu(z)} is estimated.}

\item{h}{bandwidth of the kernel.}

\item{Kernel}{name of the kernel. Possible choices are
\code{"gaussian"}, \code{"epanechnikov"}, \code{"triangular"}.}
}
\value{
A matrix of size \eqn{d \times \code{length(gridZ)}},
\eqn{\widehat{\mu}_n(z)} containing the estimated conditional means
of the \eqn{d}-dimensional random variable \eqn{X} at each point of \code{gridZ}.
}
\description{
This function estimates the conditional mean
\deqn{\mu(z) = E[X \mid Z = z]}
of a multivariate response variable using kernel smoothing.
}
\details{
The estimator is the kernel-weighted average
\deqn{
  \widehat{\mu}_n(z)
  = \sum_{i=1}^n w_{n,i}(z) X_i ,
}
where the normalized kernel weights are
\deqn{
  w_{n,i}(z)
  =
  \frac{ K\!\left( \frac{Z_i - z}{h} \right) }
       { \sum_{j=1}^n K\!\left( \frac{Z_j - z}{h} \right) }.
}
Here, \eqn{K} is the kernel function, and \eqn{h > 0} the bandwidth.
}
\examples{
# Comparison between the estimated and true conditional mean
n = 500
d = 1
Z = runif(n, -2, 2)
X = matrix(2 * Z + rnorm(n), ncol = d)
gridZ = seq(-2, 2, length.out = 50)
h = 0.3
CondMean_estimates = CondMeanEst(X, Z, gridZ, h)

true_mean = 2 * gridZ
plot(gridZ, CondMean_estimates, type = "l", col = "blue", lwd = 2,
     ylab = "Conditional Mean", xlab = "Z")
lines(gridZ, true_mean, col = "red", lwd = 2, lty = 2)
legend("topleft", legend = c("Estimated", "True"), col = c("blue", "red"),
       lty = c(1, 2), lwd = 2)

}
