% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EllCopNormalization.R
\name{DensityGenerator.check}
\alias{DensityGenerator.check}
\title{Tests the normalization of an elliptical copula generator}
\usage{
DensityGenerator.check(grid, grid_g, d, b = 1)
}
\arguments{
\item{grid}{the regularly spaced grid on which the values of the generator are given.}

\item{grid_g}{the values of the generator at points of the grid.}

\item{d}{the dimension of the space.}

\item{b}{the target value for the second identification constraint.}
}
\description{
Tests the normalization of an elliptical copula generator
}
