% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EllCopNormalization.R
\name{DensityGenerator.normalize}
\alias{DensityGenerator.normalize}
\title{Normalization of an elliptical copula generator}
\usage{
DensityGenerator.normalize(grid, grid_g, d, verbose = 0, b = 1)
}
\arguments{
\item{grid}{the regularly spaced grid on which the values of the generator are given.}

\item{grid_g}{the values of the \eqn{d}-dimensional generator at points of the grid.}

\item{d}{the dimension of the space.}

\item{verbose}{if 1, prints the estimated (alpha, beta) such that
\code{new_g(t) = alpha * old_g(beta*t)}.}

\item{b}{the target value for the second identification constraint.}
}
\description{
Normalization of an elliptical copula generator
}
\seealso{
\code{\link{EllCopSim}} for the simulation of elliptical copula samples,
\code{\link{EllCopEst}} for the estimation of elliptical copula,
\code{\link{conv_funct}} for the conversion between different representation
of the generator of an elliptical copula.
}
