% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EllDistrEstDeriv.R
\name{EllDistrDerivEst}
\alias{EllDistrDerivEst}
\title{Estimate the derivatives of a generator}
\usage{
EllDistrDerivEst(
  X,
  mu = 0,
  Sigma_m1 = diag(NCOL(X)),
  grid,
  h,
  Kernel = "gaussian",
  a = 1,
  k,
  mpfr = FALSE,
  precBits = 100,
  dopb = TRUE
)
}
\description{
Estimate the derivatives of a generator
}
\examples{

# Comparison between the estimated and true generator of the Gaussian distribution
n = 50000
d = 3
X = matrix(rnorm(n * d), ncol = d)
grid = seq(0, 5, by = 0.1)
a = 1.5

gprimeEst = EllDistrDerivEst(X = X, grid = grid, a = a, h = 0.09, k = 1)
plot(grid, gprimeEst, type = "l")

# Computation of true values
g = exp(-grid/2)/(2*pi)^{3/2}
gprime = (-1/2) *exp(-grid/2)/(2*pi)^{3/2}

lines(grid, gprime, col = "red")

}
\author{
Victor Ryan, Alexis Derumigny
}
