% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EllDistrEstDeriv.R
\name{estimate.derivatives.generator}
\alias{estimate.derivatives.generator}
\title{Estimate a generator and its derivatives}
\usage{
estimate.derivatives.generator(X, grid, d, a, h, kmax)
}
\value{
a matrix of size \code{length(grid) * (kmax + 1)}
with the estimated value of the generator and all its derivatives
at all orders until and including \code{kmax}, at all points of the grid.
}
\description{
Estimate a generator and its derivatives
}
\examples{
n = 800000
d = 3
X = matrix(rnorm(n * d), ncol = d)
grid = seq(0, 5, by = 0.1)
a = 2

estim_g_derivs = estimate.derivatives.generator(
  X = X, grid = grid, a = a, h = 0.08, kmax = 1, d = d)

# Computation of true values
g = exp(-grid/2)/(2*pi)^{3/2}
gprime = (-1/2) *exp(-grid/2)/(2*pi)^{3/2}

plot(grid, g, col = "blue", type = "l")
lines(grid, estim_g_derivs[,1])

plot(grid, gprime, col = "blue", type = "l")
lines(grid, estim_g_derivs[,2])


}
\author{
Victor Ryan, Alexis Derumigny
}
