% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EllDistrSim.R
\name{simulation_MH}
\alias{simulation_MH}
\title{Simulation of univariate random variables using the Metropolis-Hasting algorithm}
\usage{
simulation_MH(n, densityFUN, niter = 100, startValue = 1)
}
\arguments{
\item{n}{number of observations.}

\item{densityFUN}{a function giving the univariate density to sample from.}

\item{niter}{number of iterations of the Metropolis-Hasting algorithm.}

\item{startValue}{the starting value of algorithm.}
}
\value{
a vector of size n of simulated observations.
}
\description{
Simulation of univariate random variables using the Metropolis-Hasting algorithm
}
\examples{
simulation_MH(n = 10, densityFUN = function(x){exp(-x^2)})

}
