% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EllDistrEst-adaptive.R
\name{EllDistrEst.adapt}
\alias{EllDistrEst.adapt}
\title{Estimation of the generator of the elliptical distribution by kernel smoothing
with adaptive choice of the bandwidth}
\usage{
EllDistrEst.adapt(
  X,
  mu = 0,
  Sigma_m1 = diag(NCOL(X)),
  grid,
  h_firstStep,
  grid_a = NULL,
  Kernel = "gaussian",
  mpfr = FALSE,
  precBits = 100,
  dopb = TRUE
)
}
\arguments{
\item{X}{dataset, assumed to be i.i.d. observations (rows) of an elliptical
distribution.}

\item{mu}{(estimated) mean of X.}

\item{Sigma_m1}{(estimated) inverse of the covariance matrix of X.}

\item{grid}{vector containing the values at which we want the generator to be
estimated.}

\item{h_firstStep}{a vector of size \code{2} containing first-step bandwidths
to be used. The first one is used for the estimation of the asymptotic mean-squared
error. The second one is used for the first step estimation of \eqn{g}.
From these two estimators, a final value of the bandwidth \eqn{h} is determined,
which is used for the final estimator of \eqn{g}.

If \code{h_firstStep} is of length \code{1}, its value is reused for both purposes
(estimation of the AMSE and first-step estimation of \eqn{g}).}

\item{grid_a}{the grid of possible values of \code{a} to be used.
If missing, a default sequence is used.}

\item{Kernel}{name of the kernel. Possible choices are
\code{"gaussian"}, \code{"epanechnikov"}, \code{"triangular"}.}

\item{mpfr, precBits}{\code{mpfr} is a Boolean value indicating whether to use
the package \link[Rmpfr]{Rmpfr} for more precise computations. If so,
then \code{precBits} indicates how many bits of precision are used.}

\item{dopb}{a Boolean value indicating whether to print a progress bar.}
}
\value{
a list with the following elements: \itemize{
\item \code{g} a vector of size \code{n1 = length(grid)}.
Each component of this vector is an estimator of \eqn{g(x[i])}
where \code{x[i]} is the \eqn{i}-th element of the grid.

\item \code{best_a} a vector of the same size as \code{grid} indicating
for each value of the grid what is the optimal choice of \eqn{a} found by
our algorithm (which is used to estimate \eqn{g}).

\item \code{best_h} a vector of the same size as \code{grid} indicating
for each value of the grid what is the optimal choice of \eqn{h} found by
our algorithm (which is used to estimate \eqn{g}).

\item \code{first_step_g} first step estimator of \code{g}, computed using
the tuning parameters \code{best_a} and \code{h_firstStep[2]}.

\item \code{AMSE_estimated} an estimator of the part of the asymptotic MSE
that only depends on \eqn{a}.
}
}
\description{
Estimation of the generator of the elliptical distribution by kernel smoothing
with adaptive choice of the bandwidth
}
\examples{
n = 500
d = 3
X = matrix(rnorm(n * d), ncol = d)
grid = seq(0, 5, by = 0.1)

result = EllDistrEst.adapt(X = X, grid = grid, h = 0.05)
plot(grid, result$g, type = "l")
lines(grid, result$first_step_g, col = "blue")

# Computation of true values
g = exp(-grid/2)/(2*pi)^{3/2}
lines(grid, g, type = "l", col = "red")

plot(grid, result$best_a, type = "l", col = "red")
plot(grid, result$best_h, type = "l", col = "red")

sum((g - result$g)^2, na.rm = TRUE) < sum((g - result$first_step_g)^2, na.rm = TRUE)

}
\author{
Alexis Derumigny, Victor Ryan
}
