% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EllDistrSim.R
\name{EllDistrSim}
\alias{EllDistrSim}
\title{Simulation of elliptically symmetric random variables}
\usage{
EllDistrSim(n, d, A = diag(d), mu = 0, density_R2, niter = 500)
}
\arguments{
\item{n}{number of observations.}

\item{d}{dimension of X.}

\item{A}{square-root of the covariance matrix of X.}

\item{mu}{mean of X. It should be a vector of size d.}

\item{density_R2}{density of the random variable \eqn{R^2},
i.e. the density of the \eqn{||X||_2^2} if \eqn{\mu=0}
and \eqn{A} is the identity matrix.
Note that this function must return \eqn{0} for negative inputs.}

\item{niter}{number of iterations of the Metropolis-Hasting algorithm.}
}
\description{
This function uses the decomposition \eqn{X = \mu + R * A * U}
where \eqn{\mu} is the mean of X, R is the random radius,
A is the square-root of the covariance matrix of X,
and U is a uniform random variable of the d-dimensional unit sphere.
Note that R is generated using the Metropolis-Hasting algorithm.
}
\examples{
# Sample from a 3-dimensional normal distribution
X=EllDistrSim(n = 200, d = 3, density_R2 = function(x){stats::pchisq(q=x,df=3)})
plot(X[,1], X[,2])


}
\seealso{
\code{\link{EllCopSim}} for the simulation of elliptical copula samples,
\code{\link{EllCopEst}} for the estimation of elliptical distributions.
}
