% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EllDistrSim.R
\name{EllDistrSim}
\alias{EllDistrSim}
\title{Simulation of elliptically symmetric random vectors}
\usage{
EllDistrSim(
  n,
  d,
  A = diag(d),
  Sigma = NULL,
  mu = 0,
  density_R2,
  genR = list(method = "pinv")
)
}
\arguments{
\item{n}{number of observations.}

\item{d}{dimension of \eqn{X}.}

\item{A, Sigma}{A is the square-root of the covariance matrix \code{Sigma} of
\eqn{X}. It can be computed as \code{A = t(chol(Sigma))}.
For convenience, it is possible to specify \code{Sigma} directly; this
overrides \code{A}.}

\item{mu}{mean of \eqn{X}. It should be a vector of size \code{d}.}

\item{density_R2}{density of the random variable \eqn{R^2},
i.e. the density of the \eqn{||X||_2^2} if \eqn{\mu=0}
and \eqn{A} is the identity matrix.

Note that this function \strong{must} return \eqn{0} for negative inputs,
otherwise negative values of \eqn{R^2} may be generated.
The simplest way to do this is to add \verb{ * (x > 0)} at the end of the
return value of the provided \code{density_R2} function (see example below).}

\item{genR}{additional arguments for the generation of the squared radius.
It must be a list with a component method: \itemize{
\item If \code{genR$method == "pinv"}, the radius is generated
using the function \code{\link[Runuran:pinv.new]{Runuran::pinv.new()}}.
\item If \code{genR$method == "MH"},
the generation is done using the Metropolis-Hasting algorithm,
with a \eqn{N(0,1)} move at each step.
}}
}
\value{
a matrix of dimensions \code{(n, d)} of simulated observations.
}
\description{
This function uses the decomposition \eqn{X = \mu + R * A * U}
where \eqn{\mu} is the mean of \eqn{X}, \eqn{R} is the random radius,
\eqn{A} is the square-root of the covariance matrix of \eqn{X},
and \eqn{U} is a uniform random variable of the d-dimensional unit sphere.
Note that \eqn{R} is generated using the Metropolis-Hasting algorithm.
}
\examples{
# Sample from the 3-dimensional standard normal distribution
X = EllDistrSim(n = 200, d = 3, density_R2 = function(x){stats::dchisq(x=x,df=3)})
plot(X[,1], X[,2])
hist(X[,1])
cov(X)

X = EllDistrSim(n = 200, d = 3, density_R2 = function(x){stats::dchisq(x=x,df=3)},
                genR = list(method = "MH", niter = 500))
plot(X[,1], X[,2])


# Sample from the t distribution with 6 degrees of freedom

df = 6
d = 2
cov1 = rbind(c(1, 0.5), c(0.5, 1))

density_R2_student_t <- function(r2){
  s_d = 2 * pi^(d / 2) / gamma(d / 2)  # Surface area of the unit ball in R^d

  return ( (r2 > 0) * r2^(d/2 - 1) * (s_d / 2) *
    gamma((df + d)/2) / ( gamma(df/2) * gamma(1/2)^d * (df - 2) ) *
    (1 + r2 / (df - 2) )^(-(df + d)/2)
  )
}

X = EllDistrSim(n = 1000, d = d,
                Sigma = cov1, mu = c(2, 6),
                density_R2 = density_R2_student_t)
hist(X[,1])
plot(X[,1], X[,2])
cov(X)

}
\seealso{
\code{\link{EllCopSim}} for the simulation of elliptical copula samples,
\code{\link{EllCopEst}} for the estimation of elliptical distributions,
\code{\link{EllDistrSimCond}} for the conditional simulation of
elliptically distributed random vectors given some observe components.
}
